windows: Fix compilation

+++ src/logging.c
@@ -21,7 +21,7 @@ bio_logging_init(void) {
 
 	int depth = options.current_depth_in_project + 1;
 	if (options.current_filename != NULL) {
-		int len = strlen(current_filename);
+		int len = (int)strlen(current_filename);
 		int i = len - 1;
 		for (; i >= 0; --i) {
 			char ch = current_filename[i];
@@ -119,7 +119,7 @@ bio_log(
 	...
 ) {
 	filename = filename != NULL ? filename : "<unknown>";
-	int filename_len = strlen(filename);
+	int filename_len = (int)strlen(filename);
 
 	// Shorten filename by common prefix
 	if (
